+++ codelists/tests/test_hierarchy.py
@@ -1,56 +1,4 @@
-from codelists.hierarchy import Hierarchy
-
-
-def build_small_hierarchy():
-    r"""Return hierarchy with this structure:
-
-        a
-       / \
-      b   c
-     / \ / \
-    d   e   f
-    """
-
-    edges = [
-        ("a", "b"),
-        ("a", "c"),
-        ("b", "d"),
-        ("b", "e"),
-        ("c", "e"),
-        ("c", "f"),
-    ]
-
-    return Hierarchy("a", edges)
-
-
-def build_hierarchy():
-    r"""Return hierarchy with this structure:
-
-           a
-          / \
-         b   c
-        / \ / \
-       d   e   f
-      / \ / \ / \
-     g   h   i   j
-    """
-
-    edges = [
-        ("a", "b"),
-        ("a", "c"),
-        ("b", "d"),
-        ("b", "e"),
-        ("c", "e"),
-        ("c", "f"),
-        ("d", "g"),
-        ("d", "h"),
-        ("e", "h"),
-        ("e", "i"),
-        ("f", "i"),
-        ("f", "j"),
-    ]
-
-    return Hierarchy("a", edges)
+from .helpers import build_small_hierarchy
 
 
 def test_nodes():
